Clean description pragmas using AST transformation instead of in lexer
With the previous implementation, the lexer would "throw away" some
input characters without emitting any matching tokens. This causes
issues for code which assumes that there is a one-to-one matching
between input text ranges and tokens, such as the ProB 2 UI's B editor
syntax highlighter.

The new implementation performs the description cleaning on the AST,
which is okay because the AST doesn't need to correspond exactly to the
input text.