Remove handling of prob.home system property in FileSearchPathProvider
ProB 2 now always sets prob.stdlib automatically based on prob.home, so
the parser doesn't need to take care of this.

In most cases prob.home is unset anyway: prob.home is only set by the
user to use an alternative probcli with ProB 2 (usually during
development); it is not set when using the probcli bundled with ProB 2,
or when the parser is used from probcli.