Remove no longer needed case in BOM handling in Utils.readFile
The Windows/Unix differences described in the comments come from
differences in the default charset: on Unix systems it's almost always
UTF-8, but on Windows it's usually a legacy codepage like Windows-1252.
When decoding a UTF-8 BOM as Windows-1252, it results in the three
Unicode characters listed in the second if condition. This can no
longer happen, because readFile now always uses UTF-8 and no longer
relies on the default charset.