Remove support for translating *from* LaTeX in UnicodeTranslator
Recognizing LaTeX commands in the input caused some valid B code to not
translate correctly, because a backslash could stand for either the set
minus operator or the start of a LaTeX command. Translating *from*
LaTeX doesn't seem to be used anywhere currently (the Jupyter kernel
translates *to* LaTeX, but not the other way around), so it should be
safe to remove. If it's needed again in the future, a separate parser
probably needs to be implemented for inputs in LaTeX format.