Don't recognize middle dot as Event-B identifier part
In older versions of the Unicode standard, the middle dot was
considered an identifier part. This has been changed in newer versions,
but Java's Character.isUnicodeIdentifierPart still recognizes it for
compatibility. In the context of Event-B/Rodin it's obviously not an
identifier part - otherwise code such as ∃x·x=1 wouldn't parse as
expected.