Add UnicodeTranslator.toRodinUnicode for Rodin-compatible translation
For now this only removes type comments. It might be useful to add
other transformations in the future. This also gives us more
flexibility with changing the normal Unicode translation, because it no
longer needs to stay compatible with what the Rodin parser supports.
For example, the relation inverse operator could be translated to
superscript -1 instead of a Unicode tilde (the Rodin parser only
understands the latter).