Remove Helpers.parseMachineAndGetPrologOutput
It does the same thing as .getMachineAsPrologTerm, except that it
converts parse exceptions to Prolog terms first, which isn't very
useful in tests. If the exception is expected, it should be explicitly
caught and checked, and if it's not expected, it should be thrown so it
causes the test to fail.