Throw exceptions from RulesUtil methods instead of returning as terms
In most cases, the expected behavior is that the code can be parsed
successfully. If an error is actually expected, it's better to catch it
explicitly. Most tests already included a check that no exception was
returned, but some tests were missing this check even though no error
was expected, leading to undetected failures.