Replace duplicate rules file loading code in RulesMachineFilesTest
The RulesMachineFilesTest.getRulesMachineAsPrologTerm code was very
similar to RulesUtil.getFileAsPrologTerm, and the former can be
implemented based on the latter.

The helper method now cannot be used anymore in one special case
(trying to load a nonexistant file), but this is only needed in one
single test and can be easily implemented by hand.