Fix positions unintentionally not being output for some nodes
This check is supposed to suppress only nodes that have no position
information at all. Previously it was enough to check just that the node
ID and file number are missing. Now that node IDs are not assigned
anymore in most cases, this causes false positives for nodes that have
no file number, but do have line/column information.