Remove duplicated machine file lookup code from ReferencedMachines
The paths looked up through this code were only used when reporting
cycle errors, so the lookup is now performed in the error reporting code
instead. This allows reusing the existing file lookup code in
RecursiveMachineLoader instead of a second copy in ReferencedMachines.

MachineReference.path is now only set if a /*@file ...*/ pragma is
attached to the machine reference. This matches the behavior from before
the cycle detection code was added. (Some existing code also still
assumes this - e. g. the error messages in the 3-arg MachineReference
constructor only make sense in the context of file pragmas.)