Remove ReferencedMachines.siblingTable
It was identical to .referencesTable, except that some of the reference
objects contained slightly different nodes. This was apparently only
needed so that the cycle error code could detect the reference type
based on the node class. Now that the reference type is stored
explicitly, this is no longer needed, and we can just use
.referencesTable everywhere.