Remove redundant directory check from MachineReference
Specifying a directory inside a /*@file ...*/ pragma will already cause
an error once the parser tries to read the file, so there's no need to
explicitly check this case beforehand. The check was also incorrect
sometimes - specifically, if the pragma contained a relative path, this
check would incorrectly look up the path inside the working directory
instead of the machine directory (and the two are usually not
identical).