Replace ReferencedMachines.referencesTable map with a list
The map was only used for its values - nothing was ever looked up by its
key. With this change, multiple machine references with the same machine
name no longer overwrite each other. This is relevant now that
MachineReference can also store the renamed name of a machine, because
the same machine may be referenced multiple times with different renamed
names.