remove functor attribute from PrologTerm and fix bug
in equals of CompoundTerm (atom “A” was considered equal to variable “A”).
Note: the atom “[]” is no longer considered to be equal to the empty list;
but then the equals method of ListPrologTerm did not consider itself to
be equal to the atom “[]” either; and comparing longer lists with compound
terms was not done either.

Signed-off-by: Michael Leuschel <leuschel@uni-duesseldorf.de>