Change FastReadWriter to use binary instead of text output
This is important when non-ASCII bytes need to be written, e. g. for
large compound terms. Such bytes will most likely be mis-encoded when
output through a text-based Writer.

This change also requires restructuring the output handling in
CliBParser a bit, to allow outputting either text or binary data
depending on whether fastrw output is selected.