Deprecate EventBParser.getSourcePositions
This should not be needed, because all SableCC nodes and tokens extend
PositionedNode, which already provides methods to get position info.
Most likely this method is a leftover from when we extended SableCC
using aspects and manual source patches. Unfortunately we cannot remove
it yet, because Camille still uses it.