Implement a better solution for correct positions in LTL propositions
The starting position for the parser can now be overridden, which allows
passing the current position of the LTL parser on to the B parser.

This also replaces the previous code for correcting the positions of
standalone formulas (due to the implicitly added #FORMULA, etc.
prefixes). That was previously handled using the offset functionality of
ClassicalPositionPrinter and the BCompoundException.withLinesOneOff
method.