Remove redundant checking of assignment LHSes in IdentListCheck
The LHSes of assignment and operation call substitutions are specified
in the grammar as assign_list, which can only contain identifiers,
function calls, and record fields. Thus, the parser can never produce an
AST that fails this check.

The LHS of an assignment in a "substitution predicate" can contain
arbitrary expressions, even though it should have the same restrictions.
This can't be enforced in the current grammar, because of reduce/reduce
conflicts - but IdentListCheck didn't check that case either, so I don't
care.