Simplify PreLexer nesting logic and remove incorrect comment
The example from the comment is incorrect, because every parenthesis is
lexed as its own token and is handled separately before this code. The
logic of the code is also faulty - TBeginNesting and TEndNesting are two
different token classes, so a single token cannot be an instance of both
classes!