remove uses of FORCE in rule bodies
For infinitely many counter examples this causes enumeration warnings. However, it is very difficult for a user to find out the origin of these problems and fix them. Instead, FORCE is now applied in prob_java if the set is known to be finite. Otherwise the enumeration warning is handled.