Remove line/column numbers from parse exception messages
The error positions are also stored in a structured form in the
exceptions. Most users of the parser (prob_prolog, the ProB Java API,
ProB 2 UI, etc.) display this structured position info to the user, so
the line/column in the message is duplicate information.

With this change, the position info is *not* shown anymore when simply
printing a parse exception. We should probably do something about that
to help with debugging and for other users of the parser that don't
handle the structured position info.