Remove special code for local variables shadowing DEFINITIONS
This behavior is not supported by ProB - it replaces DEFINITION
identifiers regardless of context. It is also not supported by tools
that expand DEFINITIONS using text replacement (e. g. Atelier B).
Sometimes it is even expected that DEFINITIONS will be expanded inside
local variable declarations, e. g. when a DEFINITION's right-hand side
is a list of identifiers.