Remove duplicate rule for expression with description pragma
expression_in_par already falls through to expression_top_with_opt_desc,
which also allows a description pragma. The only effect of this change
is that adding a description pragma to a ; or || expression requires
an extra set of parentheses. This was also the case previously, before
version 2.13.4 (before commit f5e2c11d46dedbbedf8d213758aaa88c10815a54),
so I assume this is unlikely to break anything.