Remove double printing of exception stack trace
If the exception is thrown, the caller is responsible for dealing with
it and displaying it to the user appropriately. If the calling code
swallows the stack trace, that must be fixed in the caller, not here.
Otherwise, the same exception will get printed multiple times for
callers that do handle it correctly.